+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityTableModelTest.java
@@ -85,7 +85,7 @@ public final class SwingEntityTableModelTest extends AbstractEntityTableModelTes
 		SwingEntityTableModel employeeTableModel = createTableModel(Employee.TYPE, connectionProvider());
 		assertEquals(0, employeeTableModel.rowCount());
 		Entity accounting = connectionProvider().connection().selectSingle(Department.ID.equalTo(10));
-		employeeTableModel.conditionModel().setInConditionValues(Employee.DEPARTMENT_FK, singletonList(accounting));
+		employeeTableModel.conditionModel().setInOperands(Employee.DEPARTMENT_FK, singletonList(accounting));
 		employeeTableModel.refresh();
 		assertEquals(7, employeeTableModel.rowCount());
 	}